+++ test/integration/tests/tcti/abrmd/policyauthorize.sh
@@ -48,25 +48,28 @@ file_verifying_key_ctx="verifying_key_ctx"
 file_policyref="policyref"
 
 cleanup() {
-  rm -f  $file_pcr_value $file_policy $file_session_file $file_private_key \
+    rm -f  $file_pcr_value $file_policy $file_session_file $file_private_key \
     $file_public_key $file_verifying_key_public $file_verifying_key_name \
     $file_verifying_key_ctx $file_policyref $file_authorized_policy_1 \
     $file_authorized_policy_2
 
-  tpm2_flushcontext -S $file_session_file 2>/dev/null || true
+    tpm2_flushcontext -S $file_session_file 2>/dev/null || true
+
+    if [ "${1}" != "no-shutdown" ]; then
+        shut_down
+    fi
 }
 trap cleanup EXIT
 
 start_up
 
-cleanup
+cleanup "no-shutdown"
 
 generate_policy_authorize () {
-
-  tpm2_startauthsession -Q -S $file_session_file
-  tpm2_policyauthorize -Q -S $file_session_file  -o $3 -f $1 -q $2 -n $4
-  tpm2_flushcontext -S $file_session_file
-  rm $file_session_file
+    tpm2_startauthsession -Q -S $file_session_file
+    tpm2_policyauthorize -Q -S $file_session_file  -o $3 -i $1 -q $2 -n $4
+    tpm2_flushcontext -S $file_session_file
+    rm $file_session_file
 }
 
 openssl genrsa -out $file_private_key 2048 2>/dev/null
@@ -78,7 +81,7 @@ dd if=/dev/urandom of=$file_policyref bs=1 count=32 2>/dev/null
 
 tpm2_pcrlist -Q -L ${alg_pcr_policy}:${pcr_ids} -o $file_pcr_value
 tpm2_startauthsession -Q -S $file_session_file
-tpm2_policypcr -Q -S $file_session_file -L ${alg_pcr_policy}:${pcr_ids} -F $file_pcr_value -f $file_policy
+tpm2_policypcr -Q -S $file_session_file -L ${alg_pcr_policy}:${pcr_ids} -F $file_pcr_value -o $file_policy
 tpm2_flushcontext -S $file_session_file
 rm $file_session_file
 
@@ -90,7 +93,7 @@ tpm2_pcrextend  \
 
 tpm2_pcrlist -Q -L ${alg_pcr_policy}:${pcr_ids} -o $file_pcr_value
 tpm2_startauthsession -Q -S $file_session_file
-tpm2_policypcr -Q -S $file_session_file -L ${alg_pcr_policy}:${pcr_ids} -F $file_pcr_value -f $file_policy
+tpm2_policypcr -Q -S $file_session_file -L ${alg_pcr_policy}:${pcr_ids} -F $file_pcr_value -o $file_policy
 tpm2_flushcontext -S $file_session_file
 rm $file_session_file